+++ src/test/converter2/issues/gh2509.ts
@@ -1,0 +1,9 @@
+export interface Int {
+    /** Cb */
+    cb: () => Promise<any>;
+
+    nested: {
+        /** Cb2 */
+        cb: () => any;
+    };
+}