+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 -   Constructed references to enum types will be properly linked with `@interface`, #2508.
+-   Comments on property-methods will no longer be duplicated in generated documentation, #2509.
 -   Reduced rendered docs size by writing icons to a referenced SVG asset, #2505.
     For TypeDoc's docs, this reduced the rendered documentation size by ~30%.
 -   The HTML docs now attempt to reduce repaints caused by dynamically loading the navigation, #2491.